+++ Qwerty.ECS.Runtime/EcsTypeManager.cs
@@ -5,16 +5,16 @@ namespace Qwerty.ECS.Runtime
 {
 	internal static class EcsTypeManager
 	{
-		public static byte typeCount => m_typeCount;
-		internal static readonly int[] Sizes = new int[byte.MaxValue];
+		public static short typeCount => m_typeCount;
+		internal static readonly int[] Sizes = new int[short.MaxValue];
 
-		private static byte m_typeCount;
+		private static short m_typeCount;
 		
-		internal static byte Register<TC>() where TC : struct
+		internal static short Register<TC>() where TC : struct
 		{
 			int index = m_typeCount++;
 			Sizes[index] = Unsafe.SizeOf<TC>();
-			return (byte) index;
+			return (short) index;
 		}
 	}
 }
